feat: 实现智能体咨询会话接口与服务，含数据模型、API路由、消息持久化
- 新增 ConsultationSession 和 ConsultationMessage 数据模型，使用 BigInteger ID
- 实现会话创建、列表、详情、删除 API，支持分页和权限控制
- 更新 /api/consult/chat 接口，支持 session_id 参数，AI 回复持久化
- 添加 consultation_service.py 服务层，处理会话和消息管理
- 改进数据库初始化异常处理
- 更新 OpenAPI 规范和实现计划文档
- 删除过时的测试文件

Refs: 智能体咨询会话接口与服务实现计划

+++ app/models/consultation.py
@@ -1,12 +1,12 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey, Text
+from sqlalchemy import Column, BigInteger, String, DateTime, Boolean, ForeignKey, Text
 from sqlalchemy.orm import relationship
 from datetime import datetime, timezone
 from app.models import Base
 
 class ConsultationSession(Base):
     __tablename__ = "consultation_sessions"
-    id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     topic = Column(String(255), nullable=True)
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
     is_archived = Column(Boolean, default=False)
@@ -15,8 +15,8 @@ class ConsultationSession(Base):
 
 class ConsultationMessage(Base):
     __tablename__ = "consultation_messages"
-    id = Column(Integer, primary_key=True, index=True)
-    session_id = Column(Integer, ForeignKey("consultation_sessions.id"), nullable=False, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    session_id = Column(BigInteger, ForeignKey("consultation_sessions.id"), nullable=False, index=True)
     role = Column(String(32), nullable=False)  # 枚举，user/ai
     content = Column(Text, nullable=False)  # Markdown 格式内容
     sources = Column(Text, nullable=True)  # (RAG 引用来源: [{title, id, score}])
